+++ cppad/local/div_op.hpp
@@ -46,8 +46,8 @@ inline void forward_divvv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 
 	// Taylor coefficients corresponding to arguments and result
@@ -91,8 +91,8 @@ inline void forward_divvv_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 
 	// Taylor coefficients corresponding to arguments and result
 	Base* x = taylor + arg[0] * nc_taylor;
@@ -130,8 +130,8 @@ inline void reverse_divvv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
@@ -190,7 +190,7 @@ inline void forward_divpv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivpvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivpvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 
 	// Taylor coefficients corresponding to arguments and result
@@ -240,7 +240,7 @@ inline void forward_divpv_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivpvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivpvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 
 	// Paraemter value
 	Base x = parameter[ arg[0] ];
@@ -279,7 +279,7 @@ inline void reverse_divpv_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvvOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvvOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[1] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[1]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
 
@@ -337,7 +337,7 @@ inline void forward_divvp_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvpOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvpOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 
 	// Taylor coefficients corresponding to arguments and result
@@ -377,7 +377,7 @@ inline void forward_divvp_op_0(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvpOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvpOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
 
 	// Parameter value
 	Base y = parameter[ arg[1] ];
@@ -416,7 +416,7 @@ inline void reverse_divvp_op(
 	// check assumptions
 	CPPAD_ASSERT_UNKNOWN( NumArg(DivvpOp) == 2 );
 	CPPAD_ASSERT_UNKNOWN( NumRes(DivvpOp) == 1 );
-	CPPAD_ASSERT_UNKNOWN( arg[0] < i_z );
+	CPPAD_ASSERT_UNKNOWN( size_t(arg[0]) < i_z );
 	CPPAD_ASSERT_UNKNOWN( d < nc_taylor );
 	CPPAD_ASSERT_UNKNOWN( d < nc_partial );
